Fixed #90 - QSO fields from prev QSOs when Prefix - Callsign - Suffix
QSO fields are filled based on information from prev QSOs. If QLog
looks for a match then it begins with a full callsign. If the full
callsign is not found then it tries to find a based callsign.

Full set of fields are taken only in case when QLog finds an exact
match for full callsign. Otherwise, it takes only Name, Notes,
URL and email.