Added undocumented param "namespace"
Undocumented param used for debugging. A developer can run QLog with
specific namespace. This helps in cases where it is possible
to simultaneously run a test/develop and production versions of QLog.

The parameter changes the Application name. It causes that all runtime
files, settings, passwords and DB file are created in a different directory/namespace.

However, it remains necessary only one instance of QLog to run at a time.

NOTE: This is not a preparation for the ability to run separate databases.
It's just to make it easier for developers and testers.