Fixed #139 - CW speed synchronisation
CW Speed is synchronized in case when connected key
is not MorseOverCAT and operator enabled this feature in Rig Setting

this synchronization can be used if, for example, Winkey is connected
to the rig and the rig (eg K3X) has a build-in key.
Then an operator can begin QSO via winkey or QLog and finish it via  build-in key
in the same CW Key speed.