Fix issue with Condition Class
the initial idea was to create a singleton
for Propagation Condition and use it everywhere where it is needed.
Unfortunately, I don't know why, but this approach works everywhere
except the Windows platform.

Under Windows: On shutdown, QLog never shuts down and remains
blocked in the Propagation Conditions destructor. It was tested
that if the nam->get(URL) call is removed from the update function,
everything worked fines. I have no idea why this is so, that's why
I made it this way.