Propagation: Added HTTP Request resend mechanism
The mechanism was implemented only because of a issue with prop.kc2g.com
This site has IPv4 and IPv6 DNS record, and if the notebook is IPv4 only, QT uses an IPv6
address for the first attempt and an IPv4 address for the second attempt.
This resulted in a long interval before information was obtained from this server.