Fixed #188: Online Map incorrect Bounds if double-click on Bandmap callsign
QLog sends the position of the DXCC entity and then the exact position of
the station. In online maps, the leaflet.fitbounds function is used for each
of these changes in the Bounding box. The change is accompanied by a move animation.

Currently, it seems that if the position change comes earlier than the animation
is finished, the bounding box change is not performed properly.

Therefore the animation is turned off in QLog now.