Fixed #254 - LoTW QSL import reports unmatched QSOs sometime
QLog has a time tolerance of +-1 hour for QSL import (eQSL and LoTW).
However, if there is the same QSO in this +-1 hour interval, QLog declares
this QSO as Unmatched, because it is not able to pair them exactly
(even though they have the same times).

It seems that LoTW time tolerance is +-30 minutes. Qlog is changed
to the same interval.