Reviewed usage of QVariant isNull
Based on QT6 spec, all isNull calls had to be reviewed.

https://doc.qt.io/qt-6/qtcore-changes-qt6.html

QT6 spec:
QVariant has been rewritten to use QMetaType for all of its
operations. This implies behavior changes in a few methods:

QVariant::isNull() now only returns true if the QVariant
is empty or contains a nullptr. In Qt 5, it also returned
true for classes in qtbase which had an isNull method
themselves if that one returned true. Code relying on the old behavior
needs to check whether the contained value returns isNull – however such
code is unlikely to occur in practice, as isNull() is rarely the property
one is interested in (compare QString::isEmpty() / isNull()
and QTime::isValid / isNull).