Fixed #206 - Network notifications: IP Addresses are not resolved
Only a validator for entering the IP address was added. The hostname is
not supported due to the need to have the fastest possible processing.
If DNS is used, the first request is very slow and it completely block
the main thread. Due to the fact that the IP lists can be long, it is
not possible to implement it now. If it starts to bother, it will
have to be completely reworked.