Fixed #346 - Callbook Name field unification
QLog no longer uses the a field NICK as a source for the name field.
Instead, NAME_FMT is used in the case of QRZ and ADR_NAMR in the case of HamQTH. NICK is no longer displayed.

I tested to append NICK, but it was too long to display it.
In general, there is inconsistency in setting names and nick fields in callbook services. It can't be done right.