CWConsole: Added Word/Whole mode
CW text can be transmitted either word by word or as a whole text.

In FLDigi Keyer, a modification has been introduced
so that the end of transmission is now simulated by
pressing the "Enter" key. The Enter character is
automatically inserted upon pressing "new line"
in the console widget..