WSJTX: Removed qsl_sent lotw_qsl_sent eqsl_qsl_sent from ADIF message
the values have default values ​​defined in ADIF.
These are set to N. However, WSJTX and other apps do not send
these fields, which means that all records are set to N - Do Not Send.
It is unacceptable. So this piece of code deletes these defaults so that
in NewContact these values ​​are set from the values ​​in the GUI.

There is a risk that some clone will start sending these values ​​as well.
In this case, it will have to be handled differently. Currently it is OK